Move pyam-related computations to .compat.pyam.computations
