+++ genno/compat/pyam/computations.py
@@ -1,0 +1,32 @@
+import pyam
+import genno.computations
+
+
+def concat(*args, **kwargs):
+    """Concatenate *args*, which must all be :class:`pyam.IamDataFrame`."""
+    if isinstance(args[0], pyam.IamDataFrame):
+        # pyam.concat() takes an iterable of args
+        return pyam.concat(args, **kwargs)
+    else:
+        # genno.computations.concat() takes a variable number of positional arguments
+        return genno.computations.concat(*args, **kwargs)
+
+
+def write_report(quantity, path):
+    """Write the report identified by *key* to the file at *path*.
+
+    If *quantity* is a :class:`pyam.IamDataFrame` and *path* ends with '.csv' or
+    '.xlsx', use :mod:`pyam` methods to write the file to CSV or Excel format,
+    respectively. Otherwise, equivalent to :func:`genno.computations.write_report`.
+    """
+    if not isinstance(quantity, pyam.IamDataFrame):
+        return genno.computations.write_report(quantity, path)
+
+    if path.suffix == ".csv":
+        quantity.to_csv(path)
+    elif path.suffix == ".xlsx":
+        quantity.to_excel(path, merge_cells=False)
+    else:
+        raise ValueError(
+            f"pyam.IamDataFrame can be written to .csv or .xlsx, not {path.suffix}"
+        )
